Moved database.yml to database.yml_template
In general, you do not want your database.yml file to be stored in
source control. There was no production password in here yet, so
it wasn't a big deal, but I'm moving it now since it'll have to be
done eventually.

The deploy process (Capistrano, pushing to Heroku, etc.) should take
care of creating a proper database.yml file, which keeps the production
password out of the source code.