Results page fix for "top-4" highlighting.
* Fixed the logic that was trying to determine who was in the top-4
(order by points was sorting nil to the top of the list).

* Moved the "who is to advance" query out of the loop, since we only
need to load that list once.

* Fixed the CSS that was causing the "advance" class to only be
associated with "odd" table rows.

Note: There is a bug / missing-feature still hiding in here, which
pertains to how a tie for 4th place should be handled. The code
currently makes no attempt to handle this case.